ESC-598 rename WriteableUndertaking to UndertakingCreate and move into separate file
