+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/models/UndertakingCreate.scala
@@ -1,0 +1,41 @@
+/*
+ * Copyright 2022 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.eusubsidycompliancefrontend.models
+
+import play.api.libs.json.{Json, OFormat}
+import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.Sector.Sector
+import uk.gov.hmrc.eusubsidycompliancefrontend.models.types.{UndertakingName, UndertakingRef}
+
+case class UndertakingCreate(
+  name: UndertakingName,
+  industrySector: Sector,
+  undertakingBusinessEntity: List[BusinessEntity]
+) {
+
+  def toUndertakingWithRef(ref: UndertakingRef): Undertaking = Undertaking(
+    reference = ref,
+    name = name,
+    industrySector = industrySector,
+    industrySectorLimit = None,
+    lastSubsidyUsageUpdt = None,
+    undertakingBusinessEntity = undertakingBusinessEntity
+  )
+}
+
+object UndertakingCreate {
+  implicit val writeableUndertakingFormat: OFormat[UndertakingCreate] = Json.format[UndertakingCreate]
+}
