+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/models/json/eis/package.scala
@@ -158,14 +158,14 @@ package object eis {
 
   // convenience reads so we can store a created undertaking
   // TODO - confirm that this change is ok
-  val undertakingRequestReads: Reads[WriteableUndertaking] = new Reads[WriteableUndertaking] {
-    override def reads(json: JsValue): JsResult[WriteableUndertaking] = {
+  val undertakingRequestReads: Reads[UndertakingCreate] = new Reads[UndertakingCreate] {
+    override def reads(json: JsValue): JsResult[UndertakingCreate] = {
       val businessEntity: BusinessEntity = BusinessEntity(
         (json \ "createUndertakingRequest" \ "requestDetail" \ "businessEntity" \ "idValue").as[EORI],
         leadEORI = true
       )
       JsSuccess(
-        WriteableUndertaking(
+        UndertakingCreate(
           (json \ "createUndertakingRequest" \ "requestDetail" \ "undertakingName").as[UndertakingName],
           (json \ "createUndertakingRequest" \ "requestDetail" \ "industrySector").as[Sector],
           List(businessEntity)
