+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/services/EscService.scala
@@ -39,7 +39,7 @@ class EscService @Inject() (
   undertakingCache: UndertakingCache
 )(implicit ec: ExecutionContext) {
 
-  def createUndertaking(undertaking: WriteableUndertaking)(implicit hc: HeaderCarrier, eori: EORI): Future[UndertakingRef] =
+  def createUndertaking(undertaking: UndertakingCreate)(implicit hc: HeaderCarrier, eori: EORI): Future[UndertakingRef] =
     escConnector
       .createUndertaking(undertaking)
       .flatMap { response =>
